Further modifications for drop interpretation:
Modified getStrand and setStrand in FeatureSet to return the strand which
has been set on the FeatureSet if it contains no SeqFeatures

Added getStrand and setStrand to FeatureView

Modified TestLoader to create an empty AnnotationSet for strands with no
Annotations.

Modified StrandedZoomableApolloPanel to set strands on its two annotation
views.