FeatureEditorDialog draft class and supporting classes.
This is mainly to decide what kind of interface we want for this editor. The
editors for each type of property are intentionally inconsistant, so we
can see which one we like best.

For more detailed information, see accompanying apollo list posting.


FeatureEditorDialog - Modal dialog that contains fields to edit non-list
                      feature properties and a JTabbedPane containing
                      panels for editing list properties (like synonyms)

EvidenceEditPanel - Panel that allows editing of a list of evidence (currently
                    implemented with JTable)

CommentEditPanel - Panel that allows editing of a list of comments (currently
                   implemented with an interactive JList thingy, responds
	           to keystrokes)

DbXrefEditPanel - Panel that allows editing of a list of dbxrefs (currently
	          implemented with JTable)

SynonymEditPanel - Panel that allows editing of a list of synonyms (currently
	           implemented with an interactive JList thingy, responds
	           to unfocus events)