Added methods to FeatureSetI and supporting classes:
FeaturePairI findFeaturePairByName(String hname);

Gene findGeneByName(String name);

Both of these methods traverse FeatureSets and FeaturePairs to find the
requested feature. The implementation is identical for both FeatureSet
and DrawableFeatureSet.

Also modified Gene so that getName() returns getIdentifier().getName() . This
change was made because FeaturePairs that contained a Gene were getting
the name "no_name" instead of the name from the Gene's identifier. This change
may not be appropriate if FeaturePairs never contain Genes.