dded FlexibleFeatureTierManager to replace ExpandedFeatureTierManager - can expand and contract and show/hide individual tiers Added options to show/hide/expand/contract features in TierManager. Added sorting to EvidencePanel - the ids with the top scoring hit are shown first then they are ordered by the query start coordinate Added a FeatureProperty panel with a colour chooser for the type colours and checkboxes for visibility and expansion. Also added a TypePanel which takes a FeatureType object and makes PRopertyPanels for each property.  We have a problem with this panel.  We really need it to apply to both FeatureViews in the ApolloPanel but they have separate TierManagers and FeatureTypes. I'll leave this until Steve has returned as I don't want to generate too many conflicts.
27/08/00

When a new gene is created then a unique name for that session is
generated (Gene1, Gene2 etc).  Transcripts are then named
Gene1:1 Gene1:2 etc.  Exons are then named Gene1.exon:1
Gene1.exon:2.

GenericAnnotationSet has set/getName and set/getDesciption overridden methods
so the Identifier gets updated.

setName is overridden in Gene so the Transcript and exon names are changed
accordingly

The way unique exons are made has to be changed - previously we were comparing
identifiers to generate unique exons.  This won't work as when a new transcript
comes in to a gene we don't give it the same identifier.  We have to generate
unique exons based on coordinates.