REmoved comment from Controller.
Jeez this is slow using 1.1.8 and the corba server.  It takes
a couple of seconds to ping events to the Controller and out again.
This is really not good.  Also there is something screwed with the translations
I generate for the corba sequences - I haven't got time to look at it now
as I've a plane to catch - Steve - I'll sort this out when I get back.