Fixed vertical scrollbar positioning bugs. This involved adding a new TierManagerEvent which notifies the View when the Tier layout has changed.
Fixed horizontal scrollbar bugs - centre was being reset to
zero in setScrollValues when it shouldn't be.

Fixed problems with unknown types display and tier management.

Added setVisible and isVisible methods to ViewI to set and determine
whether a View is currently visible. I've changed the view menu
set forward/reverse visible to use this.

Modified TestLoader.Load() to throw a DAE instead of exitting. This meant
modifying all the classes which used it in their test main functions to
catch this error. The reason for doing this was to allow ApolloFrame to
be displayed even if the initial Load failed.

Fixed display update bugs.

Added code to Config class to read a configuration file from
APOLLO_ROOT/data/apollo.cfg. If this file doesn't exist default values are
set for the configuration options.

Added more gui dependencies to Makefile