Improved FeatureTreePanel functionality. Modified to use SeqFeatureI s as the DMTNs again (instead of Strings). The reason is its quicker to compare references than Strings. Needed to add toString to ExonHolder, GenericAnnotationSet and AnnotatedFeatureSet (for the JTree to use). Strands without any Annotations are not shown. If there are no annotations "No annotations" is displayed as the only node. Feature selections in the tree jump to the correct position in the display - may be whether this happens should be configurable.
Added numChildFeatures method to FeatureSet. This counts the total number of
SeqFeatures (NOT including FeatureSets) below this FeatureSet. It is important
to be able to determine if there are no SeqFeatures under a particular FeatureSet,
because many of the values returned by a FeatureSet are meaningless
if there are no SeqFeatures under it - maybe there should be a hasChildFeatures()
method which returns true if there are any - it would be quicker. I have not
added numChildFeatures to FeatureSetI yet and its only implemented in FeatureSet
ie. not for the DrawableFeatureSet.

Tidied up StrandedZoomableApolloPanel jbInit() method. There is currently a bug
with loading new data into apollo. The View limits do not get updated correctly.
I am in the process of fixing this.

Replaced LEFT and RIGHT Strings for Scrollbar positioning with ViewI.LEFTSIDE
and ViewI.RIGHTSIDE.

Added ability to start without command line arguments.