Making progress on mergeTranscripts in AnnotationView. Now at least the actions happen even if there not all the correct actions! - not all the old transcripts disappear. Partly due to getTranscript being wrong in ExonHolder, also the DrawableSeqFeature was being used in ExonI comparisons rather than its getFeature.
Modified definitions of vertical, horizontal and diagonal moves -see
AnnotationView for explanation.

Added two debug methods to FeatureChangeEvent - getTypeAsString and
getSubTypeAsString to convert the type numbers to Strings.

Overrode handleTierManagerEvent in AnnotationView because relayout is
necessary if the number of tiers in the AnnotationView changes.

Added check for if start > end in DrawableFeatureSet and
DrawableAnnotatedFeatureSet draw methods. If this occurs end is
incremented. This stops the stubby lines sticking out of the left side of
drawn features. Michele will this cause problems elsewhere?