Added setController and getController to Config.
Modified FeatureTypes to extend Observable, and created TypesObserver
class and TypesChangedEvent and TypesChangedListener to respond to
changes to the Config FeatureTypes object through the Controller.
Only adds are currently detected, and only a dummy handleTypesChangedEvent
exists in StrandedZoomableApolloPanel.

Modified all classes which created a Controller object to use the
Config.getController method instead.