Rewrote apollo startup script to do the right thing for different jdk versions. Also provided a consistent method for specifying exactly which version of the jdk the startup script will use.
What it does:

First checks to see if $APOLLO_ROOT/bin/.javapathconfig exists. If it
does exist, the contents of the file are used as the path to the java
installation. If the file doesn't exist, the path to java is
determined by analyzing the results of "which java". If the path still
can't be determined, fail.

Then, ask the version of java found above for it's version
number. Different versions are handled differently:

JDK1.1	- The lib/classes.zip file is added to the classpath, so that
          installations that have been imperfectly configured can
          still start

JDK1.2  - The Xbootclasspath parameter is specified so jacorb
          overrides the broken Java2 ORB classes

others  - No special handling occurs

The return code is the return code from java.

Notes:

* Classes are loaded using the -classpath parameter, NOT the $CLASSPATH
  variable. $CLASSPATH is entirely ignored.

* No environment variables need to be set except for $APOLLO_ROOT

* This will fail if the user doesn't have write access to their home
  directory

* If a file exists in the user's home directory named
  .javaversion.tmp, it will be erased by the startup script (this
  happens because I can't figure out how to pipe stderr in Korn shell)