Added loading and saving dialogs, activated save menus. Also did a few bug fixes, added splash screens, and modified startup procedure.
ApolloFrame:
--------------
ApolloFrame can now be initialized with NO command line parameters. If no
parameters are specified, a load region dialog box pops up. If the parameters
are specified, Apollo stars from the command line parameters, and no dialog
appears. The load/save dialog box caches which regions have been loaded
and saved in a history file, currently located at
$APOLLO_ROOT/data/apollo.use.history

Also, added lame splash screen with loading status information. This should
eventually be changed into a cool graphic of some kind.

DataAdapterChoiceDialog:
------------------------
The GUI control for loads and saves. It pops up a modal dialog box. Once
the box is closed, the getAdapter method will return a properly initialized
dataadapter.

Sequence:
---------
It turns out Java serialization cannot serialize large strings, because
it uses a crummy UTC format that stores the string length at the beginning.
So, added code to serialize the string as a char array, and to deserialize
the char array properly.

*Adapter, CurationSet:
----------------------
Added getRegion method to adapters and CurationSet. This is done to for
completeness, and so that saved serial files know what their origin region
was. In the future, a loaded serial file's region could be checked against
the requested region.

AdapterRegistry:
----------------
Changed the way CORBAAdapter is initialized.

Config:
-------
Modified to look for AutosaveFile property in apollo.cfg . AutosaveFile
specifies the path to the autosave file. If the path is not specified,
apollo.bak is the default.

TestLoader:
-----------
Lots of changes to allow the startup datasource to be specified by the new GUI
control.

FileMenu:
---------
Removed "Load" and "Load File..." options, implemented "Save" and
"Save File..." ("Save File" is more like "Save As" now). If the data was
initially loaded from a read only datasource, "Save" starts out disabled