Added Translation Start and End handling to Transcript and fixed various Translation related bugs in Exon, Transcript and ResidueProperties.
CORBAAdapter changed to set Translation Start and End - NOTE there seems to
be an off by one error in the CORBA server which is currently corrected for
in the CORBAAdapter.

DummySequenceAdapter now produces Ts and As in blocks of 10 instead of just
Ts - this is more useful for debugging.

Added base position box to StatusBar. There were problems with StatusBar
changes causing complete relayout of the display. These have been fixed.

Added siteDrag handling code in AnnotationView for Start codons and
start splices - still work to do on this.