Tidied up the draw method DrawableSeqFeature based classes. There are now four methods which can be overiden:
drawSelected()
drawUnselected()
addHighlights()
addDecorations()

or draw() can be overiden.

There are now two new classes DrawableResultSeqFeature and
DrawableResultFeatureSet which hold results (they look the same as the old
DrawableSeqFeature and DrawableFeatureSet respectively).

DrawableSeqFeature now just draws a basic rectangle feature, with no sequence
or other decorations.

DrawableFeatureSet just draws a line from start to end of the set. No line
is drawn if there is only 1 feature in the set.

Fixed bug in edge highlighting (wrong edge was highlighted when reverse
complemented).

Sites are now drawn differently (like other results but with no sequence).

Selection boxes are now two pixels wide for both results and annotations.

Also added kludge to allow sites to work in EvidencePanel - they are not in
individual FeatureSets (this would use a lot of memory) so I've created
EvidenceSelectionSet which checks for site types in the selected features
and creates temporary FeatureSets to hold each site in findSets().