Work on sequence reading for results:
Added getAllSequences to CORBAAdapter. This takes the results FeatureSet
and loads all the hit sequences for the contained FeaturePairs. It only
loads each hit sequence once and only gets the AnnotatedRegion once.

Changed DrawableResultSeqFeature to not try to load sequence if getRefSequence
is non null even if getSequence is null (in this case there is a reference
sequence but for some reason (eg. too short) the sequence for the particular
feature cannot be got. It is pointless to try reloading the sequence in this
case - infact loading sequences here is very slow and has been replaced
by the getAllSequences call above when the CurationSet is loaded - this
has the disadvantage that it uses quite a lot of memory. Very preliminary
code is in place to show the sequence.

Modified length check in getSequence in SeqFeature (was off by one).