Added scrollHack bit to ScrollableTierView. The scrollHack allows the top featureView to have the same scrolling behavior as the bottom view.
Replaced drawString(AttributedCharacterIterator, int, int) method in
SeqAlignPanel. Without that method, the code will not compile
in JDK1.2 and higher. With the method, the code will not compile under JDK1.1,
but at least it will run. Since it has to be incompatible with one of them,
I thought it should be compatible with the faster compiler.