Added JFLAGS to Makefile. This is used to set java version specific flags.
Added apollo/ver1.1 hierarchy. This contains Java 1.1 specific source
files. The only file there currently is AttributedCharacterIterator which
is missing in 1.1, but needed for the FastTranslatedGraphics implementation
To get successful compilation in 1.1 apollo/ver1.1 must be in the CLASSPATH
The Makefile contains a JFLAGS with a -classpath option to accomplish this.
This is a bit of a hack but it gets round the lack of conditional
compilation in Java