Changes for multi type tiers and also for pixel mask drawing accelerations
Tier changes
============
The type for each feature is now made up of program:dbname when data is read
using the CORBAAdapter.

The types.dat file has been modified for these types (see notes in types.dat)

tiers.dat has been added which contains the type groups in each tier.

FeatureGroupTypes and FeatureGroupProperty are derived from FeatureTypes
and FeatureProperty respectively. This allows for minimal changes to the
rest of the code.

FeatureGroupTypes reads tiers.dat and creates a set of FeatureGroupProperties
which hold the multi type information for each tier.

Config now sets up a FeatureGroupTypes and GroupTypes can be set in the
apollo.cfg file.

The DrawableSeqFeature now holds both the groupProperty and basicProperty for
each feature - the basic property is that from types.dat - this is needed
for getSize() to do size by score

Pixel Masking
=============
Added code to only draw one Feature at each pixel. This is to speed up
drawing the SiteViews. It currently only gives a 2-3 times speed up but
I'm still working on it.

As part of this I have modified the getVisibleFeatures to return a Vector of
Vectors of features (one Vector for each row).