Added ApolloNameAdapterI interface. This interface's only responsibility is to name new annotations. The ApolloNameAdapter is installed via a NameAdapterInstall directive. Only one name adapter may be installed per configuration file; multiple NameAdapterInstall directives are ignored.
The installed name adapter is registered with the same AdapterRegistry as the
other data adapters. When the parent component of the AnnotationEditor is set,
the AnnotationEditor ask's that component's parent frame for the name
adapter. After that, the name adapter is used to name every new gene and
transcript.

Because there is a new interface for this naming operation, the feature naming
methods of the existing data adapters and ApolloDataAdapterI have been removed.

FlyNameAdapter is an implementation of ApolloNameAdapterI. Right now
FlyNameAdapter uses a lame, counter-based method of naming features. Once
a good database-centric system is established, FlyNameAdapter will be updated
to take advantage of it.