Added some debug print statements.
SeqFeatureI now has a constant for unknown name (although I think you
should be able to set this most times). SeqFeature now defaults its name
to this.

ApolloFrame now takes an ApolloDataAdapterI in the constructor, which can
be supplied by DataLoader.getStartupAdapter. This allows the curationSet
passed in to be made programatically else-where, in which case the
DataLoader would have a null getStartupAdapter.

Monkeyed arround with Config.java trying to track down the cause of a
FileNotFoundException - Apparently I was trying to load suzi's types.