Got display text and display boxes working, including disabling colour text/b&w text group when text isn't displayed.
Added code into JalAction to register enablers and disablers - these are
other JalToggleActions, the state of which determine whether this JalAction is
enabled or disabled. These will be used to produce the correct interactions
between the colour schemes and PID theshold etc.

Also started making the XML command reading code optional - there's now
a CommandAdapter base class with XMLCommandAdapter and TextCommandAdapter.
TextCommandAdapter is not finished.