EnsCGIAdapter now gets sequence using dumpview CGI script on Ensembl web site.  Added a DataStream constructor to the FastaFile.
Modified ControlledObjectI so can control whether listeners are automatically
removed from closing Windows - this was to get the TypePanel displayed in
the Windows menu. This required changes to all classes which implement the
ControlledObjectI interface.

Added configuration options to set the number of bases per pixel at which
start and stop codons are first shown - changed the code to use an absolute
base/pixel value rather than a zoom factor.

Added configuration option to set initial main panel orientation and
now use an Action to control the menu item which sets this.

Added code to flash the selected type in the TypePanel when a new selection
is made - the selected type flashs a darker colour for 0.75 seconds.

Added a GuideLine which is just a vertical line which can be moved in the
main panel.

Fixed bug in FindPanel when sequence list is updated after a selection has
been made.

The FeatureTreePanel can be accessed again now in a separate window.

Several options added to ViewMenu to control the above features.

Modified the DataLoader getSequence to check whether the CurationSet has
a RegionSequence set before trying to get the region sequence using the
DataAdapter - this is needed for the GAMEAdapter to correctly set the
sequence for the FindPanel to use.