a whole lot of changes. the most significant is a change in philosophy as to the identity of exons. while it is quite true that if two transcripts contains an exon that has the same start and the same end then this is effectively a single exon that is shared in both transcripts. However, for the purposes of editing this exon cannot be treated as a single entity. If the users withs to alter one of the transcripts they want just that transcript and its version of the exon to be alter. therefore (since apollo is an editor) all every usage of an exon will indicate that it is a distinct exon. no more exonholders, they are gone. once the editing session is completed then the shared exons can be consolidate at the end before the final save.
this solves the problem of duplicating transcripts.
this solves the problem of dragging exon edges in the exon editor

The second big change is in the drag behaviour for adding exons
and evidence in the main apollo panel. shift no longer means, add
an exon. rather it simply means drag unconstrained. exons are added
if the result feature doesn't overlap any existing exons. if it
does overlap then it is considered evidence. any number of exons
and evidence can be added at the same time. much more natural behaviour
to the users

full names of transcripts being used.

new drawing mechanisms mean you must pick up an update of tiers.dat