imported some glyphs from Ace::Graphics.  functionality is pretty much the same as it was in Ace::Graphics.  dot,ex,and triangle can all function as point glyphs.  They are drawn in the center of the feature.  Triangle also is able to behave as a normal, stretchable glyph.
There are some problems in anchored_arrow because some methods from
Ace::Sequence::Feature are needed, but have no equivalent (that I was able
to find) in Bio::Graphics.  I'll see if I can find a work
around.  Specifically, the anchored arrow will always be anchored on both
ends, looking like  |-----|, but can never be |-----> or <-----|.