Fixes to translation display in DrawableAnnotatedSeqFeature - not finished. This was completely broken.
Changed labeled from a separate flag in DrawableSeqFeature to one of the
bits in the flag byte - this makes checking for decorated features faster
and saves space. Modified setLabeled in DrawableFeatureSet to set the
labeled flag on all its features.

Optimisations to findFeatures in DrawableFeatureSet.

Added hack into FeatureSelectionEvent to specify a type of selection
to cater for the special "String" (sequence) case from the FindPanel which
lead to very slow selections. FeatureView now ignores selections of
type FeatureSelectionEvent.SEQUENCE. The FindPanel sequence match
communication definitely needs changing.

Threaded FindPanel sequence selection code, and "Load sequence" dialog
displayed for lazy loads is now done with invokeLater so it can be
called from any thread.

Other small optimisations.

Removed black as a colour for the main panel 'hatpin' in BaseFineEditor