moved determination of frame from seq_feature into sequence. all of the existing methods still work, but conceptually it makes more sense to have frame be something that you ask of a sequence. that is to say a given feature is indeed in a certain frame, but that is by virtue of the fact that its start position is at some specific offset from the beginning of its reference sequence. so it makes sense to ask a sequence "what frame is the base at position X in?" this all got started from trying to resolve all the various incorrect frame data for the various results.
also added whole genome position to the curation set. region is now
specified as a seq feature on the genome. this way we can tell where
a sequence is from and if it needs to be extended (to recover a complete
annotation) it is possible to tell where it is from in the context
of the entire genome.