Added new methods to SeqFeatureI:   public long getLocalCoordLow();   public long getLocalCoordHigh();
These methods return the same as getHigh and getLow() if there is no
difference between the coordinate space you are in and the coordinate
space you define for your children (default behavior set in SeqFeature
and DrawableSeqfeature). CurationSet overrides the default behavior
because it is in choromosome space (GAMEAdapter parses it out of map_position
element) and its children are its local space of 1 to its sequence length.
When SiteView2 queries its SeqFeatureI refSeq which is a CurationSet for
its bounds in _createSites it uses these new methods as it cares about
its coordinate system not CurationSets chromsomal position.
This fixes the bug of the non-appearing Codons.

My initial fix to this was to just have CurationSet to return 1 for getLow and
seq length for getHigh(the local coordinates it defines),
but Suzi felt that getLow and getHigh being chromosome
coordinates was actually correct and thus the new methods.
This leads to the potential for features to have children that
are defined in its coordinate space.
By the way the chromsomal coordinates currently in the xml file in
<map_position> <span> have the same start and end. If we are really
going to use those coordinates at some point that should probable be
corrected.