apollo/gui/GenericFileAdapterGUI.java: make the file selection menu wider (file names were getting truncated)
apollo/gui/ColumnApolloLayout.java: standardHeight was sometimes coming out 0,
which made the next line fail.  I'm not sure what standardHeight should be if the
calculation comes out to 0, so I tried making it 1 and this seemed to result in
good behavior, but let me know if this seems wrong!
Note that this change also should be incorporated into the trunk!