DetailInfo.getName now calls feature.getDisplayName(). SeqFeature.getDisplayName just returns getName. FeatureSet.getDisplayName returns its first childs name, unless the first child is a FeaturePair, then it returns the FeaturePairs Hname. GenericAnnotationSet overrides FeatureSets behaviour and just returns its name. FeaturePair returns its sequences display id if it has one, otherwise it returns its hit name.
This code is drawn from what was previously in DetailInfo.getName.
I am trying to make it so ensembls stuff doesnt break, which needs
the FeatureSet and FeaturePair stuff. My fear here is that this logic
is specific to ensembl's data and not to others. If this is in fact true
then this naming issue has to be dealt with in the data adapter or the
data itself. I guess I'm also wondering why isnt the FeaturePair name
and FeatureSet name given these names to begin with? In other words why
not put these names in the data or have the data adapters figure them?
Unless these solutions are general, then its fine of course.