----------------------------------------------------------------------
Current snapshot of GOET. Currently, there are some problems in the
DAG viewing/post-edit node selection mechanism. We are working on it.

Also, the current version of GOET tries to show as few roots as
possible. This means that you may be suprised at what GOET decides the
root nodes are.