Added the ability for codons to be selected. Next step would be to get them to drag. SiteCodon handles selection for SiteView. SiteView via Sites create SiteCodons for findFeatures and findDrawablesAt. If a SiteCodon gets selected (via adding to Selection) it tells its Sites to select the codon at that site. This all works with the optimized SiteView that doesnt hold DrawableSeqFeatures anymore.
MouseManager and SelectionManager are new classes that are for mouse
and selection logic that are in ApolloPanel, separating out that logic
for clarity sake. They are incomplete. MouseManager eventually should be
ApolloPanels mouse listener and handle all of its mouse logic.
SelectionManager should eventually be the controller for selection,
replacing Controller just for selection. It deals with deselecting
stuff apon getting a new selection. This was relevant for SiteView
that needed to be told to  clear its selections.

Also Im wondering about the methods findFeatures, findFeaturesOld,
and findDrawablesAt. I think they are doing similar things, and perhaps
should be merged?