more touch-up to the genbank save. a method to get and set the residue type was added to abstractsequence. that way the model is equally functional for peptide sequences. although translate should now be checking this type field in order to be consistent.
also allowing for peptides produced from annotations to be
maintained without a dependency on the presence of the original
genomic sequence. before these translations were all done
dynamically. still has a bug-if the transcript structure
or translation start are modified this needs to be nulled out
and replaced with the newer version. should ultimately improve
things because translations won't be done over and over again.