Added TranslationView which displays 3 frame translations of the genomic DNA. It also highlights the used frame for each selected feature. Needed this for the small scale synteny comparisons. Added this to StrandedZoomableApolloPanel.
SyntenyPanel now sizes its SyntenyLinkPanel more correctly.

Added config options for setting the relative size of the synteny link
panel, setting whether or not two evidence panels are displayed
in the synteny panel (one attached to each of the controllers), and
setting the scale at which the TranslationView is displayed.

Changed EnsemblDbDialog so it will work properly with EnsCGIAdapter again.

Tweaked EvidenceMenu so it doesn't need to be called from an ApolloFrame

SyntenyPanel now uses Config foreground and background colours (I couldn't
see the edge matches well with a white background).

Minor fix to SyntenyLinkPanel to stop a single wrongly stranded link being
created.

Added PHASE as a field to SetDetailPanel and modified set detail panel so
it empties its table when a zero size set is passed to it.