Fixed a memory leak. The first curation set was staying around - but only the first. The first curation set is loaded differently than subsequent curation sets (I made attempts to unify the 2 but it got a bit involved) and Im not sure why but the first load triggers a deep java memory leak bug. This bug is documented at:
http://developer.java.sun.com/developer/bugParade/bugs/4489211.html

which pretty much describes the mem leak and points to:

http://developer.java.sun.com/developer/bugParade/bugs/4352535.html

I got the impression the bug has been fixed in later version of java
but is still present in 1.3.1_02-b02. When we go to 1.4 this fix
can probably be removed. The workaround is basically to dissociate
org.bdgp.swing.widget.DataAdapterChooser from the curation set it
produces. The java memory leak happens with the popup menu in the
DataAdapterChooser which causes to the DAC to linger. The DAC has
a ref to the curation set so the curation set lingers. I avoided
going into org.bdgp and modifying the DataAdapterChooser by just
subclassing it (at 1st i made it an inner class but that was
problematic as an inner class keeps a ref to its outer class) and
having it null out its curation set after its retrieved.