oh my, where  to begin. (actually not that much)
fixed a few more bugs in the error checking code for FB
style of genbank dumping

as a part of this needed to fix a few problems in internal
vs. external (project view vs. public view). got the users
names out of the source code and got config stuff out of
data models (please don't do this, please i'm begging).
this was need to fix some bugs in the way it was currently set up.

will require changes to everyone's config files (sorry, truly)