Fixed the exceptions on style changes. On a style change the old data needs to be cleared or else the tiers get into an inconsistent state and exceptions are thrown when paint happens during the load.  Now the clearing of data happens after we know that the user input is ok (like its not a bad filename). Before it was happening before that so that it would clear even on failed loads. Then the clear was moved to after the load (in ApolloFrame) but the exceptions still happen in this case. So now they happen after we know the input/filename is good.
I tried at one point to supress the paints during load but it was disastrous.
I also tried to at one point to make the tiers consistent but that
proved to be rather involved and I decided it was easier and made more
sense to just clear out the old data before the new came in.