The place in apollo where memory usage gets the most intense is during the parse(at least for game). The previous curation set was sticking around during the parse and not getting dereferenced until the parse was done and the new curation set replaced the old in all of the curation set references. To slim down mem usage during the parse the old curation set is now derferenced and cleaned up before the parse. Apollo presently clears out all features before the parsing, so there is no reason not to do this (for the way apollo is now). This involved actively nullifying curation refs before parse time (and fixing the associated errors this brought on, i tested things and everything seems to work fine, if anyone sees any trouble with loading let me know as it may be due to these changes).
As I said this is consistent with how apollo presently operates. If,
as Vivek has suggested we want the old data set to come back on parse
failures, these changes would need to be changed (shouldnt be too
hard) and there would need to be a clean separation between the old
and new curation sets (might be non-trivial). One disadvantage to that
approach would be the extra memory usage. Also, I was thinking it
would be nice if features were visualized as they were parsed, rather
than waiting for the whole dataset to come in before we can see
anything. Theres a big gain here for big datasets coming over small
wires, where the user has to wait forever to see anything. If this
approach is taken, it is unclear if we would want to revert back to
the old data set on parsing failure, or if we would just keep up
whatever part of the dataset did suceed.

Also it would be nice if we could get the mem usage down on the xml
parsing. Im open to suggestions here. Are there any xml parsers that
use up less memory? It would be also nice to have a parser that we
could grab features from as they parsed them for the above mentioned trick.