Per bevs request, the ede now comes up on the exon selected, not at the beginning of the transcript no matter what exon is selected. Also the ede scrolls to the exon if it is selected externally, used to only go to the beginning of the transcript.
If we go to allowing multiple edes we would need to disable external
selection. Although it could just react to selections on the gene that
it cared about. In other words Im proposing that the EDE really be
gene based. It should also only bring up the gene and some surrounding
region not the whole sequence for the whole entry. This would also
have the added benefit of not having all those extra sequence lines
when looking at a one transcript gene. At the moment you always see
the number of sequence lines corresponding to the gene with the most
transcripts no matter where you are. This looks confusing for genes
with few transcripts. Also the ede is really all about focusing in on
a single gene, there is no reason to have the whole entry with all
genes in one ede (that is if we decide to go the multiple ede approach
that bev is desiring).