Added annotation menu item move exons to transcript. This moves an exon from one transcript to another. This is what merge transcripts used to do. It is only allowed if there is a full transcript selected and exons from only one other transcript to disambiguate which transcript will be the receiver and which transcript the exons come from. If you select 2 whole transcripts it should be equivalent to merging transcripts.
I added new functionality to SelectionSet to do this
checking. getFirstFullySelectedTranscript() searches for the first
transcript with all of its exons selected. The transcripts vector
contains all transcripts, whether fully or partially so it could not
be used for this purpose.

There are some issues I noticed with AnnotationEditors
mergeTranscriptWithExons method. I havent modified this method so I
believe the issues were preexisting.
Translation start and stop do not get recalculated.
Overlapping exons do not get merged but remain as overlapping
exons. In this case the moved exon still reports itself as part of its
previous transcript.