Fixed merging transcripts and moving exons to transcript when exons overlap. Heres email that discusses the matter:
So I noticed that the AnnotationChangeCoalescer.executeChanges() goes
through its change events in reverse order. Is there a reason for this?
This seems counter intuitive to me and I want to switch it to go in forward order. Otherwise if you want to put in changes where the order matters you better put what needs to happen last first. Also I noticed the events get removed as they get processed and the redraw events dont
appear to get removed. Here is the code:

   public void executeChanges() {
     coalesceChanges(null);
     AnnotationChangeEvent ace = null;
     for (int i = changes.size() - 1; i >= 0; i--) {
       AnnotationChangeEvent evt
       = (AnnotationChangeEvent) changes.elementAt(i);
       if (evt.getType() == FeatureChangeEvent.REDRAW) {
         ace = evt;
    // why is this change not removed? should it be?
       } else {
         fireAnnotationChangeEvent(evt);
         changes.removeElementAt (i);
       }
     }
     if (ace != null) {
       //        System.out.println ("Carried out " + ace.toString());
       fireAnnotationChangeEvent(ace);
     }
   }

Is it possible to receive changes in the middle of executing changes,
and is this what it is trying to deal with? And is the lack of removal of the REDRAW event intentional or an omission. So I just changed it to go through the events in order and clear out the events after its done and it appears to work fine. Is there any reason this would be problematic?


So how this came up if you are curious is that Im working on a bug in transcript merging. If exons are overlapping in the 2 transcripts the exons get messy - basically they dont really merge. The reason was that AnnotationEditor would call transcript.addExon which would do the merging, but AnnotationEditor wouldnt get the message that it was a merge instead of a plain add. So AnnotationEditor would send out add events and AnnotationView would just add the exon in even though it overlapped with another exon.

My trick for solving this was rather than trying to figure out what the transcript actually did with the exons coming in (or having AnnotationView try to figure it out) was to basically send out a message that the transcript has changed. I do this by sending out a DELETE Transcript event and then an ADD Transcript event, which forces AnnotView to get rid of its old bad transcript, and pick up the new one.
This seems easier than trying to transfer the exact details of what has changed, which just doesnt seem necasary. So I'm curious what people think about that.

So the ordering problem came in because I finally discovered it was getting the DELETE before the ADD even though I added them the other way around, and thus why I want to change it for changes where the order matters.

I also found a bug in Transcript.mergedWithExons in the course of these investigations. mergedWithExons made a big merged exon that never got added nor did the merged exons get deleted. I fixed this and renamed it addExonWithOverlapCheck(Exon).