Fixed "create exon" option on AnnotationMenu. Used to cause an apollo crash. There were several issues (I dont think it was ever fully implemented). The new features strand was not being set so both AnnotationViews would process it, and the wrong one would query the parentOkay method in DrawabaleFeatureSet, which would fail and do System.exit call. It now sets the strand so it doesnt go to the wrong AnnotView, but I also took out the System.exit call, as a stack trace is good enough. System.exit() seems a bit harsh.
In checking for overlap, FlyOverlap was querying the new features refs
ref. The new feature didnt have a ref yet as it was new, so it now
checks for null refs.

A popup warning was coming up from AnnotMenu that the edit was out of
range when it wasnt. This is because only transcripts were being
checked and create exon doesnt involve an existing transcript. I took
the check out for create exon. A check does need to go in but Im
holding off until we decide what the functionality should be - see
below.

So previously it was required that a result feature be selected. But
the functionality of the newly fully implemented create exon is that
it creates an exon from scratch with no help from the selected
feature. It uses the mouse position for the start and brings up a
popup asking how long it should be. So I took out the selected feature
requirement. It still requires that no annots be selected, which could
be taken out as well I suppose.

But this does bring up the question, should create exon use the mouse
and the inputed length, as it does now, or should it use the start and
length of a result, or do we want both options - as in 2 menu items:
"create exon from scratch" and "create exon from feature"? Im awaiting
word from a curator with an opinion on the matter.